change pos to float

+++ GSMSample_4_0_Mango/GameStateManagementSample/GameStateManagementSample/Character/Player.cs
@@ -67,8 +67,8 @@ namespace GameStateManagement.SideScrollGame
                     outmessage.Write((byte)this.currentState);
                     outmessage.Write((byte)this.lastState);
                     outmessage.Write((int)this.health);
-                    outmessage.Write((int)this.position.X);
-                    outmessage.Write((int)this.position.Y);
+                    outmessage.Write((float)this.position.X);
+                    outmessage.Write((float)this.position.Y);
                     SideScrollGame.main.client.SendMessage(outmessage, NetDeliveryMethod.Unreliable);
                 }
             }
